+++ src/command_modules/azure-cli-datalake/HISTORY.rst
@@ -3,7 +3,12 @@
 Release History
 ===============
 
-0.0.1b1 (2017-03-10)
+0.0.1b2 (2017-03-13)
++++++++++++++++++++++
+
+* Fix: Add datalake package to setup.py (#2493)
+
+0.0.1b1 (2017-03-13)
 +++++++++++++++++++++
 
 * Initial Preview Release for Azure Data Lake Store and Analytics commands